Reorganize crate items
